Fix: Cast max_tokens to integer for OpenAI/Anthropic API calls

+++ app/Services/AI/AnthropicService.php
@@ -199,8 +199,8 @@ PROMPT;
      */
     public function complete(string $prompt, array $options = []): array
     {
-        $timeout = $options['timeout'] ?? config('ai.analysis.timeout', 180);
-        $maxTokens = $options['max_tokens'] ?? config('ai.analysis.max_tokens', 4000);
+        $timeout = (int) ($options['timeout'] ?? config('ai.analysis.timeout', 180));
+        $maxTokens = (int) ($options['max_tokens'] ?? config('ai.analysis.max_tokens', 4000));
         $systemMessage = $options['system_message'] ?? 'أنت محلل متخصص في تحليل مراجعات المطاعم. أجب بـ JSON صالح فقط، بدون أي نص إضافي.';
 
         $response = Http::timeout($timeout)
